+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/alert/AlertTypeEnum.java
@@ -5,7 +5,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 import static java.util.function.Function.identity;
