+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/housekeeping/job/HousekeepingJob.java
@@ -1,0 +1,7 @@
+package org.opendatadiscovery.oddplatform.housekeeping.job;
+
+import java.sql.Connection;
+
+public interface HousekeepingJob {
+    void doHousekeeping(final Connection connection);
+}
